pyproject.toml
[tool.poetry]
name = "microsense"
version = "0.1.0"
description = "open-source project utilizing MicroPython to capture and analyze telemetry data"
authors = ["Sam Reghenzi"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
gunicorn = "^22.0.0"
bump-my-version = "^0.21.1"
opentelemetry-api = "^1.24.0"
opentelemetry-sdk = "^1.24.0"
opentelemetry-instrumentation-fastapi = "^0.45b0"
opentelemetry-instrumentation-logging = "^0.45b0"
opentelemetry-exporter-otlp = "^1.24.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
